Add condition to show reminder instructions or information if reminders are active
